Add `Close() error` method to RPC client.
